Add 'mentions to modmail'  feature (#1)
* Add 'mentions to modmail'  feature

Attempts to send valid messages that start with bot mention to modmail.

Valid messages are in text channels, have content after ping, and are
not from mods. These messages with either add to a user's current thread
or start a new. The bot will send messages to the admin channel if the
bot fails to DM the user, and respects blocked users. This commit moves
some logic for user-triggered open DMs to a new function. Theses
messages currently do not use a new logviewer message type or provide
channel metadata to the logviewer, but this should be changed in the
future when implemented in logviewer.

* nit: remove extra whitespace

* Fix edge case w/ no space between ping and content

* nit: fix commented out debug code