Remove broken mention func, replace w/ on cmd err
Replace broken mention function, replace with on_command_error().

The mention function was rendered broken when the command prefix was
changed to use pings of the bots. Considering these function was rarely
(if ever used), it has been removed. In it's place, I added a
on_command_error() function, that returns the following on
CommandNotFound:
"{config.redTick} Invalid command. If you are trying to open a modmail,
please send me a DM to begin",

The snippets command/function was also removed because it remains
unimplemented and unfinshed, leading to user confusion and orphan
code.